+++ setup.py
@@ -19,17 +19,17 @@
 from setuptools import setup
 
 setup(
-    name             = 'Flutopy',
-    version          = '1.0.0',
+    name='Flutopy',
+    version='2.0.0dev',
     # url              = 'https://github.com/cfrioux/fluto',
-    license          = 'GPLv3+',
-    description      = 'Hybrid gap-filling tool',
-    long_description = 'Hybrid gap-filling tool for metabolic models. Fluto proposes solutions that fit constraint-based (FLUx balance analysis) and graph-based (TOpology) modelings of producibility. \
+    license='GPLv3+',
+    description='Hybrid gap-filling tool',
+    long_description='Hybrid gap-filling tool for metabolic models. Fluto proposes solutions that fit constraint-based (FLUx balance analysis) and graph-based (TOpology) modelings of producibility. \
 More information on usage and troubleshooting on Github: https://github.com/cfrioux/fluto',
-    author           = 'Clemence Frioux',
-    author_email     = 'clemence.frioux@gmail.com',
-    packages         = ['flutopy'],
-    package_dir      = {'flutopy' : 'flutopy'},
-    package_data     = {'flutopy' : ['encodings/*.lp']},
-    scripts          = ['fluto.py'],
+    author='Clemence Frioux',
+    author_email='clemence.frioux@gmail.com',
+    packages=['flutopy'],
+    package_dir={'flutopy': 'flutopy'},
+    package_data={'flutopy': ['encodings/*.lp']},
+    entry_points={'console_scripts': ['fluto = flutopy:main']},
 )
